requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
playwright==1.41.0
openai>=1.0.0
greenlet>=3.0.3
supabase>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11